Extract several usecases into handleChunk
