+++ src/main/kotlin/com/daniilshevtsov/deduplication/feature/SegmentGenerator.kt
@@ -1,0 +1,6 @@
+package com.daniilshevtsov.deduplication.feature
+
+import javax.inject.Inject
+
+class SegmentGenerator @Inject constructor() {
+}
