+++ src/main/kotlin/com/daniilshevtsov/deduplication/feature/SegmentStorager.kt
@@ -1,0 +1,4 @@
+package com.daniilshevtsov.deduplication.feature
+
+class SegmentStorager {
+}
