+++ src/main/kotlin/com/daniilshevtsov/deduplication/feature/TableIndexStorager.kt
@@ -1,0 +1,4 @@
+package com.daniilshevtsov.deduplication.feature
+
+class TableIndexStorager {
+}
